Add default_params.json
If params.json doesn't exist, load default_params.json

This avoids the confusion of having a params.json that must be overwritten when cloning the repository to run on a worker node.
Instead, the repository has default_params.json (which will be used if you clone the repo and ./start.sh) but it will prefer params.json if it exists (ie. when you've used create_experiment.py to make a clone of the repo to run on an async worker).