Fewer epochs, sampling only
There are two different things going on when it comes to autoencoder-GANs.

First, you have the superresolution type networks, where the discriminator is
discriminating real images from autoencoded real images.
These don't learn a manifold of real images, they learn a manifold of pixel soup
that contains all real images (but also contains a lot of non-real stuff).

Second, you have traditional GAN generators with an encoder stuck on the front.
These actually attempt to learn the manifold of real images (but usually they
suffer from mode collapse and learn only part of the manifold).

The Magical Mystery Manifold that includes all realistic images but nothing else
is a subset of the first manifold and a superset of the second.