Hack around evaluation-during-training
When you're doing training, there is no 'fold' like there is in a
stand-alone run of eval (ie, 'fold' is not one of the options in
train_classifier.py). For now, just write the fold as "None".

Two things are really at fault here: the 'fold' concept has been
stretched too far and needs to be replaced, and the whole system
of parsing options into a **dict is getting harder to reason about.

What's better than 'fold'? There are different types of evaluation,
measuring different types of things against different datasets.
Right now the hierarchy is epoch -> fold -> statistic.

This assumes that all evaluations should be graphed per epoch, that
each evaluation will run against a limited number of folds, and that
each evaluation can include a grab-bag of different statistics.
The first and last assumptions are OK but the second one is bad.

Then there's options. That's where things really need work.